refactor: Dto 패키지 구분 및 예외 중복 코드 제거
