+++ src/main/java/com/kefa/application/service/AccountService.java
@@ -1,7 +1,7 @@
 package com.kefa.application.service;
 
-import com.kefa.api.dto.request.*;
-import com.kefa.api.dto.response.*;
+import com.kefa.api.dto.request.account.*;
+import com.kefa.api.dto.response.account.*;
 import com.kefa.application.usecase.AccountUseCase;
 import com.kefa.application.usecase.AuthenticationUseCase;
 import com.kefa.application.usecase.EmailVerificationUseCase;
