+++ src/main/java/com/kefa/common/handler/GlobalExceptionHandler.java
@@ -1,8 +1,6 @@
 package com.kefa.common.handler;
 
-import com.kefa.common.exception.AuthenticationException;
-import com.kefa.common.exception.ErrorCode;
-import com.kefa.common.exception.JwtAuthenticationException;
+import com.kefa.common.exception.*;
 import com.kefa.common.response.ApiResponse;
 import com.kefa.common.response.ErrorResponse;
 import lombok.extern.slf4j.Slf4j;
@@ -14,21 +12,23 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 @ControllerAdvice
 public class GlobalExceptionHandler {
 
-    @ExceptionHandler(AuthenticationException.class)
-    public ResponseEntity<ApiResponse<?>> handleAuthException(AuthenticationException e) {
-        log.error("AuthenticationException: {}", e.getMessage());
+    @ExceptionHandler({
+        CompanyException.class,
+        EmailException.class,
+        CipherException.class,
+        AuthenticationException.class,
+        JwtAuthenticationException.class
+    })
+    public ResponseEntity<ApiResponse<?>> handleException(CustomException e) {
+        log.error("{} {} - {}",
+            e.getClass().getSimpleName(),  // 실제 예외 클래스명으로 하기 위함
+            e.getErrorCode().name(),
+            e.getMessage()
+        );
         ErrorCode errorCode = e.getErrorCode();
         return ResponseEntity
             .status(errorCode.getStatus())
             .body(ApiResponse.error(ErrorResponse.of(errorCode)));
     }
 
-    @ExceptionHandler(JwtAuthenticationException.class)
-    public ResponseEntity<ApiResponse<?>> handleJwtException(JwtAuthenticationException e) {
-        log.error("JwtException: {}", e.getMessage());
-        ErrorCode errorCode = e.getErrorCode();
-        return ResponseEntity
-            .status(errorCode.getStatus())
-            .body(ApiResponse.error(ErrorResponse.of(errorCode)));
-    }
 }
