+++ src/main/java/com/kefa/common/response/ApiResponse.java
@@ -22,12 +22,6 @@ public class ApiResponse<T> {
             .build();
     }
 
-    public static <T> ApiResponse<T> success() {
-        return ApiResponse.<T>builder()
-            .success(true)
-            .build();
-    }
-
     public static <T> ApiResponse<T> error(ErrorResponse error) {
         return ApiResponse.<T>builder()
             .success(false)
